requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.8.0
pydantic-settings==2.3.0
aiofiles==23.1.0
httpx==0.27.0
redis==5.0.8
hiredis==2.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
python-multipart==0.0.6
slowapi==0.1.9
restrictedpython==7.1
prometheus-fastapi-instrumentator==7.0.0
structlog==24.1.0
tenacity==8.2.3
uvloop==0.20.0; sys_platform != "win32"

# File recognition
python-magic==0.4.27
pillow==10.3.0
pypdf==4.2.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.2.2
chardet==5.2.0
pydub==0.25.1

# Speech-to-text (Whisper)
openai-whisper==20231117

# Ollama client
ollama==0.2.1

# NLP routing
langdetect==1.0.9
